fix: rename mynesletter component to myaccountcontent
