--- packages/vue/src/components/templates/SfMyNewsletter/SfMyNewsletter.spec.js
@@ -1,8 +1,0 @@
-import { shallowMount } from "@vue/test-utils";
-import SfMyNewsletter from "./SfMyNewsletter.vue";
-describe("SfMyNewsletter.vue", () => {
-  it("renders a component", () => {
-    const component = shallowMount(SfMyNewsletter);
-    expect(component.classes("sf-my-newsletter")).toBe(true);
-  });
-});
