--- packages/vue/src/components/templates/SfMyNewsletter/SfMyNewsletter.stories.js
@@ -1,17 +1,0 @@
-import { SfMyNewsletter } from "@storefront-ui/vue";
-
-export default {
-  title: "Components/Templates/MyAccount/MyNewsletter",
-  component: SfMyNewsletter,
-};
-
-const Template = (args, { argTypes }) => ({
-  props: Object.keys(argTypes),
-  components: { SfMyNewsletter },
-  template: `
-  <SfMyNewsletter>
-  </SfMyNewsletter>`,
-});
-
-export const Common = Template.bind({});
-Common.args = {};
